requirements.txt
google-adk
pymongo
google-api-python-client
slack-sdk
PyGithub
fastapi
uvicorn
python-dotenv
supermemory
openai
duckduckgo-search
beautifulsoup4
requests
litellm
psycopg2-binary
sqlalchemy
